% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataexample.R
\docType{data}
\name{myData}
\alias{myData}
\title{An example dataset}
\format{
A data frame with 10 rows and 4 variables
\describe{
\item{y}{Y is the outcome variable which in this datset is the average rainfall in the Sahara Desert}
\item{x}{x is a covariate matrix and it has the variables Days of Sun, Days of Partly Sunny, Days of Clouds, and Days of Heavy Clouds}
\item{sub}{sub is the subject variable and these are 10 remote locations spread 5 miles apart throughout the desert}
}
}
\usage{
myData
}
\description{
A description of the example dataset
}
\keyword{datasets}
