% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyLinearRegressionFunction.R
\name{myLinearRegression}
\alias{myLinearRegression}
\title{Run Linear Regression Analysis for a Created Dataset}
\usage{
myLinearRegression(x, y, sub)
}
\arguments{
\item{x}{A covariate Matrix}

\item{y}{A Vector of Outcomes}

\item{sub}{A list of Subjects (integers corresponding to each row in matrix x)}
}
\value{

}
\description{
This function allows someone to create a covariate matrix with a set of outcomes and run linear regression analysis
}
\examples{
myLinearRegression(x,y,sub)
myLinearRegression(xvalues, yvalues, subjects)
}
